test: move sqlfluff specific test cases to test_sqlfluff.py
